Split code/init (forgotten)
