+++ src/scss/_menu.scss
@@ -69,7 +69,7 @@
 		}
 
 		&:hover,
-		&:cl {
+		&:focus {
 			background-color: var(--#{$prefix}light-hover-color);
 		}
